some refactor v3

+++ internal/models/models.go
@@ -19,7 +19,7 @@ type UserBalance struct {
 type UserOrder struct {
 	Number     string    `db:"order_number" json:"number"`
 	Status     string    `db:"status" json:"status"`
-	Accrual    int       `db:"accrual" json:"accrual"`
+	Accrual    float32   `db:"accrual" json:"accrual"`
 	UploadedAt time.Time `db:"uploaded_at" json:"uploaded_at"`
 }
